refactor: change file name
Changed name
- racingCarGame.js -> racingCarEventHandler.js
