+++ src/class/racingCarEventHandler.js
